+++ src/com/descabouda/model/constants/ClassConstant.scala
@@ -3,5 +3,5 @@ package com.descabouda.model.constants
 import com.descabouda.model.BaseConstant
 
 class ClassConstant extends BaseConstant {
-  var name_index: Short = null
+  var name_index: Integer = null
 }
