+++ src/com/descabouda/model/constants/StringConstant.scala
@@ -3,5 +3,5 @@ package com.descabouda.model.constants
 import com.descabouda.model.BaseConstant
 
 class StringConstant extends BaseConstant {
-  var string_index: Short = null
+  var string_index: Integer = null
 }
