+++ src/com/descabouda/model/constants/Utf8Constant.scala
@@ -3,6 +3,6 @@ package com.descabouda.model.constants
 import com.descabouda.model.BaseConstant
 
 class Utf8Constant extends BaseConstant {
-  var length: Short = null
+  var length: Integer = null
   var bytes: Array[Byte] = null
 }
